Remove use of Data::Dump::Streamer
Rob Mueller noticed that in an environment where enough inotify events
are being generated sch that the kernel event queue overflows, the
kernel can generate an overflow event, which we're not watching for and
so we end up in the "unexpected event" path.

Once there, Data::Dump::Streamer tries really hard to dump whatever
thing you gave it in a way that it can reconstruct it. The thing is the
event includes the "watcher" object, which itself has a closure over
$self, and because inotify needs to setup watchers on every subdir, when
you Data::Dump::Streamer it, you end up with this massive structure with
1000+ internal members that all reference each other being dumped. This
results in each error is attempting to dump 4M to syslog, which of
course just ignores it with a EMSGSIZE error and so it's totally lost.

Additionally, generating that structure takes noticeable time (several
seconds), thus the 100% CPU usage.

Handling the overflow event (and any other unexpected events) is
something we need to fix, but even so, error paths should be fast and
uncompolicated. So here we're just switching back to Data::Dumper. Its
already in the box, its fast, and also it won't try to look inside
closures. We can improve it later if it turns out to be wrong and/or not
good enough.